requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.10.0
click==8.1.3
Django==4.1.3
djangorestframework==3.14.0
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.2
platformdirs==2.5.4
pluggy==1.0.0
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
pytest-testdox==3.0.1
pytz==2022.6
sqlparse==0.4.3
